+++ lib/arb_utils.dart
@@ -1,9 +1,10 @@
 /// A library with tools for handling .arb files
 library arb_utils;
 
+export 'src/arb_client.dart';
 export 'src/primitives/add_missing_metadata.dart';
 export 'src/primitives/diff.dart';
 export 'src/primitives/merge.dart';
 export 'src/primitives/sort.dart';
+export 'src/utils/locale.dart';
 export 'src/utils/process_new_keys.dart';
-export 'src/arb_client.dart';